feat: implementado dto e validação para somente permitir que usuário autenticado delete a própria conta

+++ src/user/user.controller.ts
@@ -9,13 +9,15 @@ import {
   Patch,
   ParseIntPipe,
   ValidationPipe,
+  ForbiddenException,
 } from '@nestjs/common';
-import { User as UserModel } from '@prisma/client';
 import { UserService } from './user.service';
 import { AuthGuard } from 'src/auth/auth.guard';
 import { PublicUserDto } from './dto/public-user.dto';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
+import { DeleteUserDto } from './dto/delete-user.dto';
+import { CurrentUser } from 'src/auth/current-user.decorator';
 
 @Controller('user')
 export class UserController {
@@ -51,7 +53,17 @@ export class UserController {
 
   @UseGuards(AuthGuard)
   @Delete(':id')
-  async deleteUser(@Param('id', ParseIntPipe) id: number): Promise<UserModel> {
+  async deleteUser(
+    @Param('id', ParseIntPipe) id: number,
+    @Body(new ValidationPipe()) body: DeleteUserDto,
+    @CurrentUser('sub') userSub: number,
+  ): Promise<PublicUserDto> {
+    if (userSub !== id) {
+      throw new ForbiddenException('Você só pode excluir a sua própria conta.');
+    }
+
+    await this.userService.validatePasswordForId(id, body.password);
+
     return this.userService.deleteUser({ id });
   }
 }
